Struct anybar::Anybar

pub struct Anybar {
    pub port: u16,
    pub color: Option<Color>,
}

The Anybar handle.

Fields

port

The UDP Port the Anybar is connected to

color

The last color that has been set.

When no color has been set yet, color is None.

Methods

impl Anybar

fn new(port: u16) -> Anybar

Create a new Anybar instance, connected to the given UDP port.

port may be any port between 0 and 6553.

Panics

Panics if the port has a non-valid value outside of the range between 0 and 6553.

Examples

let custom_bar = Anybar::new(1708);

assert_eq!(custom_bar.port, 1708);

fn set_color(&mut self, color: Color)

Set a new color.

Panics

Panics if the UDP socket can not be bound.